pyproject.toml
[project]
name = "supertelligence-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "eddy", email = "eddy@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "discord-py (>=2.5.2,<3.0.0)",
    "openai (>=1.82.1,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "aiohttp (>=3.12.6,<4.0.0)"
]

[tool.poetry]
name = "supertelligence-backend"
version = "0.1.0"
description = ""
authors = ["eddy <eddy@example.com>"]
packages = [{include = "supertelligence_backend", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
